+++ BP2415/Commands/Applications/PingApp.cs
@@ -1,27 +1,21 @@
 using DisCatSharp.ApplicationCommands;
 using DisCatSharp.ApplicationCommands.Attributes;
 using DisCatSharp.ApplicationCommands.Context;
-using DisCatSharp.CommandsNext.Attributes;
 using DisCatSharp.Entities;
 using DisCatSharp.Enums;
 
-namespace BP2415.Commands.Application
+namespace BP2415.Commands.Applications
 {
-    public class ShutApp : ApplicationCommandsModule
+    public class PingApp : ApplicationCommandsModule
     {
-        [SlashCommand("shut", "Schaltet den Bot ab")]
-        [RequireOwner]
-        [Hidden]
+        [SlashCommand("ping", "Pingt den Bot an")]
         public async Task Ping(InteractionContext ctx)
         {
-            var bp = ctx.Client;
-
             await ctx.CreateResponseAsync(InteractionResponseType.ChannelMessageWithSource,
                 new DiscordInteractionResponseBuilder
                 {
-                    Content = "***Herunterfahren...***"
+                    Content = ":ping_pong: " + ctx.Client.Ping + "ms"
                 });
-            await bp.DisconnectAsync();
         }
     }
 }
